+++ CHANGELOG.md
@@ -11,9 +11,11 @@
     - Gabriel Alencar
  - Melhoria via Pull Request: [Code format](https://github.com/totvs/advpl-vscode/pull/420)
     - Robson Rogério Silva
+ - Melhoria: [**Compilação de pasta** não mostra nome do fonte compilado sem warnings](https://github.com/totvs/advpl-vscode/issues/436)
  - Correção: Conexão com repositorio SQLite
  - Correção: Deleção de fonte binario 19
  - Correção: [Problemas com Debug com path com acentos](https://github.com/totvs/advpl-vscode/issues/422)
+ - Correção: [Exclusão de fonte do RPO no binário 191205P](https://github.com/totvs/advpl-vscode/issues/416)
 
 ## 0.14.3
  - Melhoria: Avisar o link do uso do alpha, quando usnado o binario 19 com ssl.